add new dashboard service

+++ services/admin_service.py
@@ -1,42 +1,39 @@
-from rest_framework.exceptions import NotFound, ValidationError, APIException
+from rest_framework.exceptions import NotFound, APIException
+
 from super_admin.models import AdminUser
+
 
 class AdminService:
 
-    @staticmethod
-    def create_admin_user(validated_data):
-        if validated_data.get('password') != validated_data.pop('confirm_password', None):
-            raise ValidationError("Passwords do not match")
+  @staticmethod
+  def create_admin_user(validated_data):
+    validated_data.pop('confirm_password', None)
+    validated_data['is_staff'] = True
 
-        try:
-            user = AdminUser.objects.create_user(**validated_data)
-        except Exception as e:
-            raise APIException(f"Error creating admin user: {str(e)}")
+    try:
+      user = AdminUser.objects.create_user(**validated_data)
+    except Exception as e:
+      raise APIException(f"Error creating admin user: {str(e)}")
 
-        return user
+    return user
 
-    @staticmethod
-    def update_admin_user(user_id, validated_data):
-        try:
-            user = AdminUser.objects.get(id=user_id)
-        except AdminUser.DoesNotExist:
-            raise NotFound(f"Admin with id {user_id} not found")
-        except Exception as e:
-            raise APIException(f"Error updating admin user: {str(e)}")
+  @staticmethod
+  def update_admin_user(user_id, validated_data):
+    user = AdminUser.objects.filter(id=user_id).first()
+    if not user:
+      raise NotFound(f"Admin with id {user_id} not found")
 
-        for key, value in validated_data.items():
-            setattr(user, key, value)
-        user.save()
-        return user
+    for key, value in validated_data.items():
+      setattr(user, key, value)
 
-    @staticmethod
-    def delete_admin_user(user_id):
-        try:
-            user = AdminUser.objects.get(id=user_id)
-        except AdminUser.DoesNotExist:
-            raise NotFound(f"Admin with id {user_id} not found")
-        except Exception as e:
-            raise APIException(f"Error deleting admin user: {str(e)}")
+    user.save()
+    return user
 
-        user.delete()
-        return True
+  @staticmethod
+  def delete_admin_user(user_id):
+    user = AdminUser.objects.get(id=user_id)
+    if not user:
+      raise NotFound(f"Admin with id {user_id} not found")
+
+    user.delete()
+    return True
